Move/merge more test modules
This moves mislocated tests to their proper locations to match the files
they test, merging tests into existing modules where necessary.
