+++ tests/unit/cli/test_daemons.py
