+++ tests/unit/cloud/test_map_conf.py
